// scripts/types.ts
export interface ActivistDetail {
	id: number;
	name: string;
	description: string;
	address: string;
	contact: Contact;
	socials: Socials;
	website: string;
	tags: string[];
	logoUrl: string;
	pictures: string[];
	slug: string;
}

export interface Contact {
	email: string;
	phone: string;
}

export interface Socials {
	facebook: string;
	linkedin: string;
	instagram: string;
}

export interface EventDetail {
	id: number;
	name: string;
	description: string;
	address: string;
	contact: Contact;
	socials: Socials;
	website: string;
	tags: string[];
	logoUrl: string;
	pictures: string[];
	slug: string;
}

export interface Contact {
	email: string;
	phone: string;
}

export interface Socials {
	facebook: string;
	linkedin: string;
	instagram: string;
}

export interface OrganizationDetail {
	id: number;
	name: string;
	description: string;
	address: string;
	contact: Contact;
	socials: Socials;
	website: string;
	tags: string[];
	logoUrl: string;
	pictures: string[];
	slug: string;
}

export interface Contact {
	email: string;
	phone: string;
}

export interface Socials {
	facebook: string;
	linkedin: string;
	instagram: string;
}